[Test] RecordingScene Test 수정
